requirements.txt
langchain
PyPDF2
python-dotenv
streamlit
faiss-cpu
altair
tiktoken
langchain_openai
langchain-community
langchain-core
streamlit-authenticator
langchain-experimental
unstructured[all-docs]
unstructured-inference
psutil
poppler-utils
pillow
pydantic
lxml
tesseract
pytesseract
IPython